requirements.txt
aniso8601==9.0.1
black==22.10.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
PyJWT==1.4.2
pyodbc==4.0.34
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.42
tomli==2.0.1
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.10.0
